Prepare backprop setup to manage objects related to model learning
